# modules/apps/data-engine/data-engine-rest-impl/rest-openapi.yaml
components:
    schemas:
        DataDefinition:
            description: https://www.schema.org/DataDefinition
            properties:
                contentSpaceId:
                    format: int64
                    type: integer
                dataDefinitionFields:
                    items:
                        $ref: '#/components/schemas/DataDefinitionField'
                    type: array
                dateCreated:
                    format: date-time
                    type: string
                dateModified:
                    format: date-time
                    type: string
                description:
                    items:
                        $ref: '#/components/schemas/LocalizedValue'
                    type: array
                id:
                    format: int64
                    type: integer
                name:
                    items:
                        $ref: '#/components/schemas/LocalizedValue'
                    type: array
                storageType:
                    type: string
                userId:
                    format: int64
                    type: integer
            type: object
        DataDefinitionField:
            properties:
                defaultValue:
                    type: string
                fieldType:
                    type: string
                id:
                    format: int64
                    type: integer
                indexable:
                    type: boolean
                label:
                    items:
                        $ref: '#/components/schemas/LocalizedValue'
                    type: array
                localizable:
                    type: boolean
                name:
                    type: string
                repeatable:
                    type: boolean
                tip:
                    items:
                        $ref: '#/components/schemas/LocalizedValue'
                    type: array
            type: object
        DataLayout:
            properties:
                dataDefinitionId:
                    format: int64
                    type: integer
                dataLayoutPages:
                    items:
                        $ref: '#/components/schemas/DataLayoutPage'
                    type:
                        array
                dateCreated:
                    format: date-time
                    type: string
                dateModified:
                    format: date-time
                    type: string
                defaultLanguageId:
                    type: string
                description:
                    items:
                        $ref: '#/components/schemas/LocalizedValue'
                    type: array
                id:
                    format: int64
                    type: integer
                name:
                    items:
                        $ref: '#/components/schemas/LocalizedValue'
                    type: array
                paginationMode:
                    type: string
                userId:
                    format: int64
                    type: integer
            type: object
        DataLayoutColumn:
            properties:
                columnSize:
                    type: integer
                fieldNames:
                    items:
                        type: string
                    type:
                        array
            type: object
        DataLayoutPage:
            properties:
                dataLayoutRows:
                    items:
                        $ref: '#/components/schemas/DataLayoutRow'
                    type:
                        array
                description:
                    items:
                        $ref: '#/components/schemas/LocalizedValue'
                    type: array
                title:
                    items:
                        $ref: '#/components/schemas/LocalizedValue'
                    type: array
            type: object
        DataLayoutRow:
            properties:
                dataLayoutColums:
                    items:
                        $ref: '#/components/schemas/DataLayoutColumn'
                    type:
                        array
            type: object
        DataRecord:
            description: https://www.schema.org/DataRecord
            properties:
                dataRecordCollectionId:
                    format: int64
                    type: integer
                dataRecordValues:
                    items:
                        $ref: '#/components/schemas/DataRecordValue'
                    type: array
                id:
                    format: int64
                    type: integer
        DataRecordCollection:
            description: https://www.schema.org/DataRecordCollection
            properties:
                dataDefinitionId:
                    format: int64
                    type: integer
                description:
                    items:
                        $ref: '#/components/schemas/LocalizedValue'
                    type: array
                id:
                    format: int64
                    type: integer
                name:
                    items:
                        $ref: '#/components/schemas/LocalizedValue'
                    type: array
            type: object
        DataRecordValue:
            properties:
                key:
                    type: string
                value:
                    type: object
        LocalizedValue:
            description: https://www.schema.org/LocalizedValue
            properties:
                key:
                    type: string
                value:
                    type: string
info:
    description: ""
    title: ""
    version: v1.0
openapi: 3.0.1
paths:
    "/content-spaces/{content-space-id}/data-definitions":
        get:
            parameters:
                - in: path
                  name: content-space-id
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: keywords
                  schema:
                      type: string
                - in: query
                  name: page
                  required: true
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  required: true
                  schema:
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/DataDefinition"
                                type: array
                    description: ""
            tags: ["DataDefinition"]
        post:
            parameters:
                - in: path
                  name: content-space-id
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DataDefinition"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataDefinition"
                    description: ""
            tags: ["DataDefinition"]
    "/content-spaces/{content-space-id}/data-layout":
        get:
            parameters:
                - in: path
                  name: content-space-id
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: page
                  required: true
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  required: true
                  schema:
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/DataLayout"
                                type: array
                    description: ""
            tags: ["DataLayout"]
    "/content-spaces/{content-space-id}/data-record-collections":
        get:
            parameters:
                - in: path
                  name: content-space-id
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: keywords
                  schema:
                      type: string
                - in: query
                  name: page
                  required: true
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  required: true
                  schema:
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/DataRecordCollection"
                                type: array
                    description: ""
            tags: ["DataRecordCollection"]
    "/data-definitions/{data-definition-id}":
        delete:
            parameters:
                - in: path
                  name: data-definition-id
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                204:
                    content:
                        application/json: {}
                    description: ""
            tags: ["DataDefinition"]
        get:
            parameters:
                - in: path
                  name: data-definition-id
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataDefinition"
                    description: ""
            tags: ["DataDefinition"]
        put:
            parameters:
                - in: path
                  name: data-definition-id
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DataDefinition"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataDefinition"
                    description: ""
            tags: ["DataDefinition"]
    "/data-definitions/{data-definition-id}/data-layouts":
        post:
            parameters:
                - in: path
                  name: data-definition-id
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DataLayout"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataLayout"
                    description: ""
            tags: ["DataLayout"]
    "/data-definitions/{data-definition-id}/data-record-collections":
        get:
            parameters:
                - in: path
                  name: data-definition-id
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: keywords
                  schema:
                      type: string
                - in: query
                  name: page
                  required: true
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  required: true
                  schema:
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/DataRecordCollection"
                                type: array
                    description: ""
            tags: ["DataRecordCollection"]
        post:
            parameters:
                - in: path
                  name: data-definition-id
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DataRecordCollection"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataRecordCollection"
                    description: ""
            tags: ["DataRecordCollection"]
    "/data-layouts/{data-layout-id}":
        delete:
            parameters:
                - in: path
                  name: data-layout-id
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                204:
                    content:
                        application/json: {}
                    description: ""
            tags: ["DataLayout"]
        get:
            parameters:
                - in: path
                  name: data-layout-id
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataLayout"
                    description: ""
            tags: ["DataLayout"]
        put:
            parameters:
                - in: path
                  name: data-layout-id
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DataLayout"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataLayout"
                    description: ""
            tags: ["DataLayout"]
    "/data-record-collections/{data-record-collection-id}":
        delete:
            parameters:
                - in: path
                  name: data-record-collection-id
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                204:
                    content:
                        application/json: {}
                    description: ""
            tags: ["DataRecordCollection"]
        get:
            parameters:
                - in: path
                  name: data-record-collection-id
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataRecordCollection"
                    description: ""
            tags: ["DataRecordCollection"]
        put:
            parameters:
                - in: path
                  name: data-record-collection-id
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DataRecordCollection"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataRecordCollection"
                    description: ""
            tags: ["DataRecordCollection"]
    "/data-record-collections/{data-record-collection-id}/data-records":
        get:
            parameters:
                - in: path
                  name: data-record-collection-id
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: page
                  required: true
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  required: true
                  schema:
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/DataRecord"
                                type: array
                    description: ""
            tags: ["DataRecord"]
        post:
            parameters:
                - in: path
                  name: data-record-collection-id
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DataRecord"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataRecord"
                    description: ""
            tags: ["DataRecord"]
    "/data-records/{data-record-id}":
        delete:
            parameters:
                - in: path
                  name: data-record-id
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                204:
                    content:
                        application/json: {}
                    description: ""
            tags: ["DataRecord"]
        get:
            parameters:
                - in: path
                  name: data-record-id
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataRecord"
                    description: ""
            tags: ["DataRecord"]
        put:
            parameters:
                - in: path
                  name: data-record-id
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DataRecord"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DataRecord"
                    description: ""
            tags: ["DataRecord"]